Pull Cw721Contract struct and Maps into state.rs
