+++ contracts/cw721-base/src/contract.rs
@@ -1,10 +1,9 @@
 use serde::de::DeserializeOwned;
 use serde::Serialize;
-use std::marker::PhantomData;
 
 use cosmwasm_std::{
-    to_binary, Addr, Binary, BlockInfo, Deps, DepsMut, Env, MessageInfo, Order, Pair, Response,
-    StdError, StdResult, Storage,
+    to_binary, Binary, BlockInfo, Deps, DepsMut, Env, MessageInfo, Order, Pair, Response, StdError,
+    StdResult,
 };
 
 use cw0::maybe_addr;
@@ -14,11 +13,11 @@ use cw721::{
     Cw721Execute, Cw721Query, Cw721ReceiveMsg, Expiration, NftInfoResponse, NumTokensResponse,
     OwnerOfResponse, TokensResponse,
 };
-use cw_storage_plus::{Bound, IndexedMap, Item, Map, MultiIndex};
+use cw_storage_plus::Bound;
 
 use crate::error::ContractError;
 use crate::msg::{ExecuteMsg, InstantiateMsg, MintMsg, MinterResponse, QueryMsg};
-use crate::state::{token_owner_idx, Approval, TokenIndexes, TokenInfo};
+use crate::state::{Approval, Cw721Contract, TokenInfo};
 
 // version info for migration info
 const CONTRACT_NAME: &str = "crates.io:cw721-base";
@@ -26,74 +25,6 @@ const CONTRACT_VERSION: &str = env!("CARGO_PKG_VERSION");
 
 const DEFAULT_LIMIT: u32 = 10;
 const MAX_LIMIT: u32 = 30;
-
-pub struct Cw721Contract<'a, T, C>
-where
-    T: Serialize + DeserializeOwned + Clone,
-{
-    pub contract_info: Item<'a, ContractInfoResponse>,
-    pub minter: Item<'a, Addr>,
-    pub token_count: Item<'a, u64>,
-    /// Stored as (granter, operator) giving operator full control over granter's account
-    pub operators: Map<'a, (&'a Addr, &'a Addr), Expiration>,
-    pub tokens: IndexedMap<'a, &'a str, TokenInfo<T>, TokenIndexes<'a, T>>,
-
-    _custom_response: PhantomData<C>,
-}
-
-impl<T, C> Default for Cw721Contract<'static, T, C>
-where
-    T: Serialize + DeserializeOwned + Clone,
-    C: CustomMsg,
-{
-    fn default() -> Self {
-        Self::new(
-            "nft_info",
-            "minter",
-            "num_tokens",
-            "operators",
-            "tokens",
-            "tokens__owner",
-        )
-    }
-}
-
-impl<'a, T, C> Cw721Contract<'a, T, C>
-where
-    T: Serialize + DeserializeOwned + Clone,
-    C: CustomMsg,
-{
-    fn new(
-        contract_key: &'a str,
-        minter_key: &'a str,
-        token_count_key: &'a str,
-        operator_key: &'a str,
-        tokens_key: &'a str,
-        tokens_owner_key: &'a str,
-    ) -> Self {
-        let indexes = TokenIndexes {
-            owner: MultiIndex::new(token_owner_idx, tokens_key, tokens_owner_key),
-        };
-        Self {
-            contract_info: Item::new(contract_key),
-            minter: Item::new(minter_key),
-            token_count: Item::new(token_count_key),
-            operators: Map::new(operator_key),
-            tokens: IndexedMap::new(tokens_key, indexes),
-            _custom_response: PhantomData,
-        }
-    }
-
-    pub fn token_count(&self, storage: &dyn Storage) -> StdResult<u64> {
-        Ok(self.token_count.may_load(storage)?.unwrap_or_default())
-    }
-
-    pub fn increment_tokens(&self, storage: &mut dyn Storage) -> StdResult<u64> {
-        let val = self.token_count(storage)? + 1;
-        self.token_count.save(storage, &val)?;
-        Ok(val)
-    }
-}
 
 impl<'a, T, C> Cw721Execute<T, C> for Cw721Contract<'a, T, C>
 where
