+++ contracts/cw721-base/src/state.rs
@@ -1,10 +1,78 @@
 use schemars::JsonSchema;
-use serde::{Deserialize, Serialize};
-
-use cosmwasm_std::{Addr, BlockInfo};
-use cw721::Expiration;
-use cw_storage_plus::{Index, IndexList, MultiIndex};
 use serde::de::DeserializeOwned;
+use serde::{Deserialize, Serialize};
+use std::marker::PhantomData;
+
+use cosmwasm_std::{Addr, BlockInfo, StdResult, Storage};
+
+use cw721::{ContractInfoResponse, Expiration};
+use cw_storage_plus::{Index, IndexList, IndexedMap, Item, Map, MultiIndex};
+
+pub struct Cw721Contract<'a, T, C>
+where
+    T: Serialize + DeserializeOwned + Clone,
+{
+    pub contract_info: Item<'a, ContractInfoResponse>,
+    pub minter: Item<'a, Addr>,
+    pub token_count: Item<'a, u64>,
+    /// Stored as (granter, operator) giving operator full control over granter's account
+    pub operators: Map<'a, (&'a Addr, &'a Addr), Expiration>,
+    pub tokens: IndexedMap<'a, &'a str, TokenInfo<T>, TokenIndexes<'a, T>>,
+
+    pub(crate) _custom_response: PhantomData<C>,
+}
+
+impl<T, C> Default for Cw721Contract<'static, T, C>
+where
+    T: Serialize + DeserializeOwned + Clone,
+{
+    fn default() -> Self {
+        Self::new(
+            "nft_info",
+            "minter",
+            "num_tokens",
+            "operators",
+            "tokens",
+            "tokens__owner",
+        )
+    }
+}
+
+impl<'a, T, C> Cw721Contract<'a, T, C>
+where
+    T: Serialize + DeserializeOwned + Clone,
+{
+    fn new(
+        contract_key: &'a str,
+        minter_key: &'a str,
+        token_count_key: &'a str,
+        operator_key: &'a str,
+        tokens_key: &'a str,
+        tokens_owner_key: &'a str,
+    ) -> Self {
+        let indexes = TokenIndexes {
+            owner: MultiIndex::new(token_owner_idx, tokens_key, tokens_owner_key),
+        };
+        Self {
+            contract_info: Item::new(contract_key),
+            minter: Item::new(minter_key),
+            token_count: Item::new(token_count_key),
+            operators: Map::new(operator_key),
+            tokens: IndexedMap::new(tokens_key, indexes),
+            _custom_response: PhantomData,
+        }
+    }
+
+    pub fn token_count(&self, storage: &dyn Storage) -> StdResult<u64> {
+        Ok(self.token_count.may_load(storage)?.unwrap_or_default())
+    }
+
+    pub fn increment_tokens(&self, storage: &mut dyn Storage) -> StdResult<u64> {
+        let val = self.token_count(storage)? + 1;
+        self.token_count.save(storage, &val)?;
+        Ok(val)
+    }
+}
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct TokenInfo<T> {
